refactor: 비속어 로직 common 패키지로 이동
