+++ src/main/java/kitchenpos/common/infra/DefaultPurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.infra;
+package kitchenpos.common.infra;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
