+++ src/test/java/kitchenpos/application/FakePurgomalumClient.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.product.infra.PurgomalumClient;
+import kitchenpos.common.infra.PurgomalumClient;
 
 import java.util.Arrays;
 import java.util.List;
